+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Changed
 - For very large source images, a 2500px transform will be downloaded instead of the original when importing files from the DAM
 - Escape DAM now uses its own log file
+- Elements' canonical IDs are now recorded in the `sourceElementId` column for the `escapedam_importedfiles` database table
+- Escape DAM fields no longer returns previously imported assets (a new asset is created for every import)
 ### Fixed
 - Fixed a bug where the selected "Import Location" wasn't selected in field setting forms